+++ lib/agent0/agent0/hyperdrive/state/agent/trade_result.py
